Refactor letters filename
