+++ tests/app/letters/test_send_letter_jobs.py
@@ -8,7 +8,7 @@ from tests import create_authorization_header
 
 
 def test_send_letter_jobs(client, mocker):
-    mock_celery = mocker.patch("app.letters.send_letter_jobs.notify_celery.send_task")
+    mock_celery = mocker.patch("app.letters.rest.notify_celery.send_task")
     job_ids = {"job_ids": [str(uuid.uuid4()), str(uuid.uuid4()), str(uuid.uuid4())]}
 
     auth_header = create_authorization_header()
@@ -27,7 +27,7 @@ def test_send_letter_jobs(client, mocker):
 
 
 def test_send_letter_jobs_throws_validation_error(client, mocker):
-    mock_celery = mocker.patch("app.letters.send_letter_jobs.notify_celery.send_task")
+    mock_celery = mocker.patch("app.letters.rest.notify_celery.send_task")
 
     job_ids = {"job_ids": ["1", "2"]}
